+++ commands/setconf.js
@@ -1,4 +1,5 @@
-const db = require('../database.js');
+const Config = require('../config.json');
+const Keyv = require('keyv');
 
 module.exports = {
 	name: 'setconf',
@@ -7,9 +8,13 @@ module.exports = {
 	guildOnly: true,
 	args: true,
 	usage: '<parameter> [new value]',
-	execute(args, message) {
+	async execute(args, message) {
 		// Exits early if the message's author doesn't have the ADMINISTRATOR Permission.
 		if (!message.member.hasPermission('ADMINISTRATOR')) return message.reply('This command is only available for admins.');
+
+		// Database
+		const keyv = new Keyv(Config.db);
+		keyv.on('error', err => console.error('Connection Error', err));
 
 		// The property command.args = true,
 		// so no need to check args[0].
@@ -20,7 +25,7 @@ module.exports = {
 		if (typeof newKeyVal !== 'undefined') {
 
 			if (key === 'prefix') {
-				db.set(`prefix_${message.guild.id}`, newKeyVal);
+				await keyv.set(`prefix_${message.guild.id}`, newKeyVal);
 				message.channel.send(`My prefix has been set to: "${newKeyVal}"`);
 			}
 			else {
@@ -29,8 +34,7 @@ module.exports = {
 		}
 		// GET
 		else {
-			const value = db.get(`${key}_${message.guild.id}`);
-
+			const value = await keyv.get(`${key}_${message.guild.id}`);
 			if (value) message.channel.send(`Parameter: "${key}" = "${value}"`);
 			else message.reply(`"${key}" is not a valid parameter.`);
 		}